Change to `bqlist-lock-enable'
Though I had thought that `enable-foo' *is* convenient according to
the portion of "Emacs Lisp Coding Conventions":

  “Occasionally, for a command name intended for users to use, it is
  more convenient if some words come before the package's name prefix.”

Now I've been puzzled that it is not clear to me what are the "some
words". Even defun and defvar are prefixed these days for some
famous packages.  (`bqlist-lock-enable' is not a command name though)

So, I've changed my mind.

Signed-off-by: Takeshi Banse <takebi@laafc.net>